+++ tests/testthat/test-cube-bases.R
@@ -16,10 +16,10 @@ with_mock_crunch({
     ## Load a ton of cube fixtures via the tab book feature
     ds <- loadDataset("test ds")
     m <- multitables(ds)[[1]]
-    with_POST("https://app.crunch.io/api/datasets/1/multitables/tabbook-result.json", {
+    with_POST("https://app.crunch.io/api/datasets/1/multitables/tabbook-result/", {
         book1 <- tabBook(m, data=ds, format="json")
     })
-    with_POST("https://app.crunch.io/api/datasets/1/multitables/tabbook-array-result.json", {
+    with_POST("https://app.crunch.io/api/datasets/1/multitables/tabbook-array-result/", {
         book2 <- tabBook(m, data=ds, format="json")
     })